pyproject.toml
[tool.poetry]
name = "insurance-prediction-app"
version = "0.1.0"
description = ""
authors = ["Yannick Habecker <yannick.habecker@openknowledge.de>"]
readme = "README.md"
packages = [
    {include = "insurance_prediction/app", from = "src"},
    {include = "insurance_prediction/model", from = "src"},
    {include = "insurance_prediction/monitoring", from = "src"},
]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
tensorflow-cpu = { version = "2.12.0", markers = "platform_machine == 'x86_64' or platform_machine == 'amd64'" }
tensorflow-aarch64 = { version = "2.12.0", markers = "platform_machine == 'aarch64'" }
tensorflow-macos = { version = "2.12.0", markers = "platform_machine == 'arm64'" }
evidently = { version = "0.3.0" }
prometheus_client = { version = "0.16.0"}
fastapi = "0.95.1"
uvicorn = "0.21.1"
StrEnum = "0.4.10"
pandas = "2.0.1"
statsmodels = "0.13.5"

[tool.poetry.dev-dependencies]
pylint = "2.17.3"
pytest = "7.3.1"

# Used only for the simulate-drift demo
httpx = "0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
train = 'insurance_prediction.train:train'
validate = 'insurance_prediction.train:validate'
simulate-drift = 'insurance_prediction.demo:simulate_drift'
simulate-drift-curl = 'insurance_prediction.demo:simulate_drift_curl'
